# BE/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LingoPrep API",
    "description": "Intelligent IELTS & Aptis Speaking & Writing Practice & Assessment Platform powered by AI",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/dashboard/learner": {
      "get": {
        "tags": [
          "Dashboard & Analytics"
        ],
        "summary": "Get Learner Dashboard",
        "operationId": "get_learner_dashboard_api_dashboard_learner_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearnerDashboardStats"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/dashboard/notifications": {
      "get": {
        "tags": [
          "Dashboard & Analytics"
        ],
        "summary": "Get Notifications",
        "operationId": "get_notifications_api_dashboard_notifications_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/NotificationResponse"
                  },
                  "type": "array",
                  "title": "Response Get Notifications Api Dashboard Notifications Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/dashboard/notifications/{notification_id}/read": {
      "put": {
        "tags": [
          "Dashboard & Analytics"
        ],
        "summary": "Mark Notification Read",
        "operationId": "mark_notification_read_api_dashboard_notifications__notification_id__read_put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Notification Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExamType": {
        "type": "string",
        "enum": [
          "IELTS",
          "APTIS"
        ],
        "title": "ExamType"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LearnerDashboardStats": {
        "properties": {
          "total_submissions": {
            "type": "integer",
            "title": "Total Submissions"
          },
          "speaking_count": {
            "type": "integer",
            "title": "Speaking Count"
          },
          "writing_count": {
            "type": "integer",
            "title": "Writing Count"
          },
          "average_band": {
            "type": "number",
            "title": "Average Band"
          },
          "highest_band": {
            "type": "number",
            "title": "Highest Band"
          },
          "cefr_level": {
            "type": "string",
            "title": "Cefr Level"
          },
          "current_streak_days": {
            "type": "integer",
            "title": "Current Streak Days"
          },
          "total_practiced_minutes": {
            "type": "integer",
            "title": "Total Practiced Minutes"
          },
          "skill_radar": {
            "items": {
              "$ref": "#/components/schemas/SkillRadarItem"
            },
            "type": "array",
            "title": "Skill Radar"
          },
          "trend_history": {
            "items": {
              "$ref": "#/components/schemas/TrendPoint"
            },
            "type": "array",
            "title": "Trend History"
          },
          "recommended_questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionResponse"
            },
            "type": "array",
            "title": "Recommended Questions"
          }
        },
        "type": "object",
        "required": [
          "total_submissions",
          "speaking_count",
          "writing_count",
          "average_band",
          "highest_band",
          "cefr_level",
          "current_streak_days",
          "total_practiced_minutes",
          "skill_radar",
          "trend_history",
          "recommended_questions"
        ],
        "title": "LearnerDashboardStats"
      },
      "NotificationResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Link"
          },
          "is_read": {
            "type": "boolean",
            "title": "Is Read"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "message",
          "type",
          "is_read",
          "created_at"
        ],
        "title": "NotificationResponse"
      },
...
      "QuestionResponse": {
        "properties": {
          "exam_type": {
            "$ref": "#/components/schemas/ExamType"
          },
          "skill": {
            "$ref": "#/components/schemas/SkillType"
          },
          "part": {
            "type": "string",
            "title": "Part"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic",
            "default": "General"
          },
          "difficulty": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Difficulty",
            "default": "Medium"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "instructions": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructions"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "prep_time_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prep Time Seconds",
            "default": 60
          },
          "time_limit_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Limit Seconds",
            "default": 120
          },
          "min_words": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Words"
          },
          "max_words": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Words"
          },
          "sample_answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sample Answer"
          },
          "rubric_criteria": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rubric Criteria"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags",
            "default": []
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active",
            "default": true
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "exam_type",
          "skill",
          "part",
          "title",
          "prompt",
          "id",
          "created_at"
        ],
        "title": "QuestionResponse"
      },
...
      "SkillRadarItem": {
        "properties": {
          "skill_name": {
            "type": "string",
            "title": "Skill Name"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "max_score": {
            "type": "number",
            "title": "Max Score",
            "default": 9.0
          }
        },
        "type": "object",
        "required": [
          "skill_name",
          "score"
        ],
        "title": "SkillRadarItem"
      },
      "SkillType": {
        "type": "string",
        "enum": [
          "SPEAKING",
          "WRITING"
        ],
        "title": "SkillType"
      },
...
      "TrendPoint": {
        "properties": {
          "date": {
            "type": "string",
            "title": "Date"
          },
          "band": {
            "type": "number",
            "title": "Band"
          },
          "type": {
            "type": "string",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "date",
          "band",
          "type"
        ],
        "title": "TrendPoint"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  },
...
}